requirements.txt
alembic==1.7.7
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
Pillow==9.1.0
pip==22.1
psycopg2==2.9.3
pycparser==2.21
requests==2.27.1
setuptools==58.1.0
SQLAlchemy==1.4.36
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
